rename home component and split it into logic and layout, implement navigation component as header, fix styling errors that resulted from those changes
