+++ src/main/app/src/App.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
-import Home from './Components/HomeComponent/Home';
+import LandscapeOverview from './Components/LandscapeComponent/LandscapeOverview/LandscapeOverview';
 import Landscape from './Components/LandscapeComponent/Landscape/Landscape';
 import Man from './Components/ManComponent/Man';
+import Layout from './Components/LayoutComponent/Layout';
 
 import './App.scss';
 
@@ -11,9 +12,11 @@ const App: React.FC = () => {
   return (
     <Router>
       <Switch>
-        <Route exact path='/' component={Home} />
-        <Route exact path='/landscape/:identifier' component={Landscape} />
-        <Route exact path='/man/:usage' component={Man} />
+        <Layout>
+          <Route exact path='/' component={LandscapeOverview} />
+          <Route exact path='/landscape/:identifier' component={Landscape} />
+          <Route exact path='/man/:usage' component={Man} />
+        </Layout>
       </Switch>
     </Router>
   );
