+++ src/main/app/src/Components/EventComponent/Events.scss
