+++ src/main/app/src/Components/EventComponent/Events.test.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import Events from './Events';
+import { MemoryRouter } from 'react-router-dom';
+
+it('should render Events component', () => {
+  const { getByText } = render(
+    <MemoryRouter>
+      <Events />
+    </MemoryRouter>
+  );
+  expect(getByText('loading...')).toBeInTheDocument();
+});
