--- src/main/app/src/Components/HomeComponent/Home.scss
@@ -1,133 +1,0 @@
-@import '../../Ressources/styling/colors.scss';
-@import '../../Ressources/styling/font.scss';
-
-.homeContainer {
-  .content {
-
-    .landscapeContainer {
-
-      .MuiTypography-overline {
-        border-bottom: 1px solid $border-color;
-      }
-
-      .floatButton {
-        float: right;
-      }
-
-      .control {
-        background-color: $header-color;
-        color: $font-color;
-        text-transform: uppercase;
-        text-shadow: 0 0 10px $text-shadow-color;
-        box-shadow: 0px 0px 12px $box-shadow-color;
-        border: 1px solid $border-color;
-        cursor: pointer;
-      }
-
-      .previewItem {
-
-      }
-    }
-  }
-
-  .header {
-
-    margin-bottom: 0px;
-
-    .first {
-      background-color: $header-color;
-      border-top-left-radius: 25px;
-    }
-
-    .title {
-      font-size: $header-font-size;
-      font-family: "Roboto", "Helvetica", "Arial", sans-serif;;
-      font-weight: bold;
-      flex-grow: 1;
-      text-transform: uppercase;
-      text-align: center;
-      padding: 0;
-    }
-
-    .last {
-      background-color: $header-color;
-    }
-  }
-
-  .main {
-    margin-bottom: 0;
-    margin-top: 0;
-
-    //TODO move to upper level
-    .sidebar {
-      padding: 0px;
-
-      .spacer {
-        background-color: $header-color;
-      }
-
-      .spacer:first-child {
-        margin-bottom: 0.3rem;
-      }
-
-      .spacer:last-child {
-        margin-top: 0.3rem;
-        height: 4rem;
-      }
-    }
-
-    .elbow1 {
-      background-color: $header-color;
-    }
-
-    .elbow2 {
-      padding: 0px;
-
-      .elbow-outer {
-        background-color: $header-color;
-        width: 100%;
-        height: inherit;
-
-        .elbow-inner {
-          background-color: $background-grey;
-          width: 100%;
-          height: 2rem;
-          border-top-left-radius: 50px;
-        }
-      }
-    }
-
-    .bar {
-      margin-top: 1rem;
-      margin-bottom: 1rem;
-      height: 1rem;
-
-      .first {
-        background-color: $header-color;
-        border-top-left-radius: 25px;
-        border-bottom-left-radius: 25px;
-        padding: 0;
-      }
-
-      .title {
-        font-size: $large-font-size;
-        flex-grow: 1;
-        text-transform: uppercase;
-        text-align: center;
-        padding: 0;
-        padding-left: 0.3rem;
-        //background-color: darken($header-color, 10%);
-      }
-
-      .item {
-        background-color: $header-color;
-        padding: 0;
-      }
-
-      .no-item {
-        padding: 0;
-        border-left: 1px solid black;
-      }
-    }
-  }
-}
