+++ src/main/app/src/Components/LandscapeComponent/LandscapeOverview/LandscapeOverview.scss
@@ -1,0 +1,57 @@
+@import '../../../Ressources/styling/colors.scss';
+@import '../../../Ressources/styling/font.scss';
+
+.homeContainer {
+  .landscapeContainer {
+    .MuiTypography-overline {
+      border-bottom: 1px solid $border-color;
+    }
+
+    .floatButton {
+      float: right;
+    }
+
+    .control {
+      background-color: $header-color;
+      color: $font-color;
+      text-transform: uppercase;
+      text-shadow: 0 0 10px $text-shadow-color;
+      box-shadow: 0px 0px 12px $box-shadow-color;
+      border: 1px solid $border-color;
+      cursor: pointer;
+    }
+
+    .bar {
+      margin-top: 1rem;
+      margin-bottom: 1rem;
+      height: 1rem;
+
+      .first {
+        background-color: $header-color;
+        border-top-left-radius: 25px;
+        border-bottom-left-radius: 25px;
+        padding: 0;
+      }
+
+      .title {
+        font-size: $large-font-size;
+        flex-grow: 1;
+        text-transform: uppercase;
+        text-align: center;
+        padding: 0;
+        padding-left: 0.3rem;
+        //background-color: darken($header-color, 10%);
+      }
+
+      .item {
+        background-color: $header-color;
+        padding: 0;
+      }
+
+      .no-item {
+        padding: 0;
+        border-left: 1px solid black;
+      }
+    }
+  }
+}
