+++ src/main/app/src/Components/LayoutComponent/Layout.scss
@@ -1,0 +1,4 @@
+.content {
+  margin-top: -192px;
+  margin-left: 10vw;
+}
