+++ src/main/app/src/Components/LayoutComponent/Layout.tsx
@@ -1,0 +1,23 @@
+import React, { ReactElement } from 'react';
+
+import Navigation from '../NavigationComponent/Navigation';
+import './Layout.scss';
+
+interface Props {
+  children: string | ReactElement | ReactElement[];
+}
+
+/**
+ * Containts our site layout, Navigation on top, content below
+ * @param param0
+ */
+const Layout: React.FC<Props> = ({ children }) => {
+  return (
+    <React.Fragment>
+      <Navigation />
+      <div className='content'>{children}</div>
+    </React.Fragment>
+  );
+};
+
+export default Layout;
